short to in ldi

+++ jhakala/test_dir/src/statements/op_ldi.c
@@ -6,7 +6,7 @@
 /*   By: jhakala <jhakala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/23 11:53:48 by jhakala           #+#    #+#             */
-/*   Updated: 2020/08/06 16:49:24 by jhakala          ###   ########.fr       */
+/*   Updated: 2020/08/08 15:44:53 by jhakala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int		op_ldi(t_game *game, int place, t_carriage *c)
 	else if (types[0][0] == 2)
 		val1 = types[0][1];
 	else
-		val1 = read_types(game->arena, place + (types[0][1] % IDX_MOD), REG_SIZE);
+		val1 = read_types(game->arena, place + ((short)types[0][1] % IDX_MOD), REG_SIZE);
 	if (types[1][0] == 1)
 		val2 = c->reg[types[1][1] - 1];
 	else if (types[1][0] == 2)
